fix(main-repair): unique StringRaw stable-hash tag (was dup w/ RegExpEscape) + extract object/mod.rs tests to tests.rs (under 2000-cap)
